+++ SimuCom/de.uka.ipd.sdq.scheduler/src/de/uka/ipd/sdq/scheduler/processes/IWaitingProcess.java
@@ -6,6 +6,6 @@ public interface IWaitingProcess {
 	
 	public ISchedulableProcess getProcess();
 
-	public int getNumRequested();
+	public long getNumRequested();
 
 }
